backend/requirements.txt
django==5.2.12
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
Pillow==12.1.1
qrcode[pil]==8.2
django-filter==25.2
twilio>=8.0.0
